테스트케이스 + 불필요한 테스트케이스 제거

+++ internal/app/service/token/token_test.go
@@ -16,5 +16,5 @@ func TestLogic(t *testing.T) {
 	cl := &TokenClaim{}
 	tokenModule.ParseToken(to, cl)
 	assert.Equal(t, cl.Uuid, "uuid")
-	assert.Equal(t, cl.UserId, uint(1))
+	assert.Equal(t, cl.UserId, 1)
 }
